requirements.txt
# Core
torch>=2.3.0
transformers>=4.44.0
datasets>=2.20.0
accelerate>=0.33.0
peft>=0.12.0

# Config
hydra-core>=1.3.2
omegaconf>=2.3.0

# Data
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.13.0
requests>=2.31.0

# Semantic IDs
sentence-transformers>=3.0.0

# Experiment tracking
wandb>=0.17.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0

# Flash Attention — comment out for TPU / L4 without CUDA 11.8+
# flash-attn>=2.6.0

# QLoRA quantization — comment out for TPU
# bitsandbytes>=0.43.0
